split iface struct into many functions
